sd35-gptq-baseline/requirements.txt
# =============================================================================
# SD 3.5 Medium GPTQ Quantization Baseline - Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# Note: Install PyTorch separately first for your CUDA version
# =============================================================================

# -----------------------------------------------------------------------------
# Core Deep Learning
# -----------------------------------------------------------------------------
# PyTorch - install separately with CUDA support
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# Hugging Face ecosystem
transformers>=4.40.0
diffusers>=0.28.0
accelerate>=0.30.0
safetensors>=0.4.0

# -----------------------------------------------------------------------------
# Stable Diffusion 3.5 Specific
# -----------------------------------------------------------------------------
# T5 text encoder support
sentencepiece>=0.1.99

# -----------------------------------------------------------------------------
# Quantization
# -----------------------------------------------------------------------------
# No external quantization library needed - we implement GPTQ from scratch
# following the Q-DiT approach

# -----------------------------------------------------------------------------
# Image Processing
# -----------------------------------------------------------------------------
Pillow>=10.0.0
opencv-python>=4.8.0

# -----------------------------------------------------------------------------
# Metrics
# -----------------------------------------------------------------------------
# CLIP score - using open_clip to avoid PyTorch version issues
open-clip-torch>=2.24.0

# FID calculation
scipy>=1.11.0
torchmetrics>=1.2.0

# pytorch-fid for reference FID calculation
pytorch-fid>=0.3.0

# -----------------------------------------------------------------------------
# Data Handling
# -----------------------------------------------------------------------------
datasets>=2.16.0
numpy>=1.24.0
pandas>=2.0.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
tqdm>=4.66.0
huggingface_hub>=0.20.0
pyyaml>=6.0
jsonlines>=4.0.0

# -----------------------------------------------------------------------------
# Logging and Experiment Tracking
# -----------------------------------------------------------------------------
tensorboard>=2.15.0
# wandb>=0.16.0  # Optional: uncomment for W&B tracking

# -----------------------------------------------------------------------------
# Development
# -----------------------------------------------------------------------------
ipython>=8.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
